requirements.txt
# Core ML/AI Libraries
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
chromadb>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.38

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
datasets>=2.14.0
nltk>=3.8.0

# Web Scraping & APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.0
serpapi>=0.1.0

# Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.6.0
seaborn>=0.12.0
wordcloud>=1.9.0
networkx>=3.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0.0
